ignore

+++ LastSectionReaderWriter.cpp
@@ -8,7 +8,7 @@ void PrintCurrentSetting() {
     printf("Password Tip:%s \n", current_setting.passwordTip.data());    
     printf("Locked: %d\n", current_setting.locked);
     printf("ID: %d\n", current_setting.id);
-    for (size_t i = 0; i < current_setting.blocks.size(); i++)
+    for (uint16_t i = 0; i < current_setting.blocks.size(); i++)
     {
         printf("Block %zu: Mode:%d, Max Try:%d, Current Try:%d, Block code:%d\n", i,
                current_setting.blocks[i].mode,
@@ -23,7 +23,7 @@ void ResetCurrentSetting() {
     current_setting.passwordTip = INIT_TIP;
     current_setting.locked = LOCKED255;
     current_setting.id = (uint32_t)time(NULL);
-    for (size_t i = 0; i < current_setting.blocks.size(); i++)
+    for (uint16_t i = 0; i < current_setting.blocks.size(); i++)
     {
         current_setting.blocks[i] = {MODE255, INIT_MAX_TRY, 0, i};
     }
